# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ChargePointDetail": {
        "example": {
          "active_charging_profiles": [],
          "active_reservations": [],
          "connectors": [
            {
              "connector_id": 1,
              "error_code": "NoError",
              "status": "Charging",
              "timestamp": "2026-05-07T12:34:56+00:00"
            }
          ],
          "cp_id": "CP_BERLIN_001",
          "firmware_version": "1.2.3",
          "last_heartbeat_at": "2026-05-07T12:35:00+00:00",
          "last_status": "Charging",
          "last_status_at": "2026-05-07T12:34:56+00:00",
          "model": "X1",
          "online": true,
          "pod_id": "ocpp-gw-7f9c4-x2k8q",
          "request_id": "9b3c5d18-1f7c-4b6a-8e0e-5b9a3c4f2e10",
          "serial_number": "SN-12345",
          "vendor": "ACME"
        },
        "properties": {
          "active_charging_profiles": {
            "items": {
              "$ref": "#/components/schemas/ChargingProfile"
            },
            "title": "Active Charging Profiles",
            "type": "array"
          },
          "active_reservations": {
            "items": {
              "$ref": "#/components/schemas/Reservation"
            },
            "title": "Active Reservations",
            "type": "array"
          },
          "connectors": {
            "items": {
              "$ref": "#/components/schemas/ConnectorStatus"
            },
            "title": "Connectors",
            "type": "array"
          },
          "cp_id": {
            "description": "Stable identifier the charger sent in BootNotification.",
            "title": "Cp Id",
            "type": "string"
          },
          "firmware_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firmware Version"
          },
          "last_heartbeat_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Heartbeat At"
          },
          "last_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Last-write-wins across connectors. For multi-connector chargers, prefer `connectors[]` (the per-connector array) \u2014 `last_status` is a single-string convenience for callers that don't need per-connector resolution.",
            "title": "Last Status"
          },
          "last_status_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Status At"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "online": {
            "description": "True iff some pod currently holds the charger's WS.",
            "title": "Online",
            "type": "boolean"
          },
          "pod_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Which gateway pod owns the WS. None when offline.",
            "title": "Pod Id"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "serial_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serial Number"
          },
          "vendor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor"
          }
        },
        "required": [
          "cp_id",
          "online",
          "request_id"
        ],
        "title": "ChargePointDetail",
        "type": "object"
      },
...
      "ChargingProfile": {
        "description": "Mirror of the OCPP 1.6 SetChargingProfile request payload, as\npersisted server-side. The shape is dynamic per the OCPP spec \u2014\nwe treat it as opaque-ish for OpenAPI purposes.",
        "properties": {
          "charging_profile_id": {
            "title": "Charging Profile Id",
            "type": "integer"
          },
          "charging_profile_kind": {
            "title": "Charging Profile Kind",
            "type": "string"
          },
          "charging_profile_purpose": {
            "title": "Charging Profile Purpose",
            "type": "string"
          },
          "charging_schedule": {
            "additionalProperties": true,
            "title": "Charging Schedule",
            "type": "object"
          },
          "recurrency_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurrency Kind"
          },
          "stack_level": {
            "title": "Stack Level",
            "type": "integer"
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          }
        },
        "required": [
          "charging_profile_id",
          "stack_level",
          "charging_profile_purpose",
          "charging_profile_kind",
          "charging_schedule"
        ],
        "title": "ChargingProfile",
        "type": "object"
      },
...
      "ConnectorStatus": {
        "description": "One row from the `cp_status` ClickHouse table \u2014 the most-recent\nStatusNotification per connector.",
        "properties": {
          "connector_id": {
            "description": "0 is the charger itself; 1+ are physical connectors.",
            "minimum": 0.0,
            "title": "Connector Id",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "info": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Info"
          },
          "status": {
            "description": "OCPP 1.6 \u00a73.1.1 enum: Available, Preparing, Charging, \u2026",
            "title": "Status",
            "type": "string"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO 8601 of the StatusNotification.",
            "title": "Timestamp"
          },
          "vendor_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Error Code"
          },
          "vendor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Id"
          }
        },
        "required": [
          "connector_id",
          "status"
        ],
        "title": "ConnectorStatus",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "Stable error shape per `docs/integration/02-gateway-rest-api.md`.\n\nAll non-2xx responses from `/api/v1/*` carry this body. The 11\n`error_code` values are the closed enum from\n`src/eveys_ocpp/api/_errors.py` (BAD_REQUEST, UNAUTHORIZED,\nFORBIDDEN, UNKNOWN_CP_ID, UNKNOWN_TRANSACTION_ID,\nUNKNOWN_RESERVATION_ID, CHARGER_OFFLINE, CHARGER_TIMEOUT,\nWINDOW_TOO_LARGE, RATE_LIMITED, INTERNAL_ERROR).",
        "example": {
          "error": "unknown cp_id: CP_DOES_NOT_EXIST",
          "error_code": "UNKNOWN_CP_ID",
          "request_id": "9b3c5d18-1f7c-4b6a-8e0e-5b9a3c4f2e10"
        },
        "properties": {
          "error": {
            "description": "Human-readable description for logs.",
            "title": "Error",
            "type": "string"
          },
          "error_code": {
            "description": "Stable machine code; one of the 11 closed-enum values.",
            "title": "Error Code",
            "type": "string"
          },
          "request_id": {
            "description": "Correlation id mirroring the X-Request-ID header.",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "error",
          "error_code",
          "request_id"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeterValueSample": {
        "properties": {
          "connector_id": {
            "title": "Connector Id",
            "type": "integer"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "measurand": {
            "description": "OCPP 1.6 measurand: Energy.Active.Import.Register, \u2026",
            "title": "Measurand",
            "type": "string"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Id"
          },
          "unit": {
            "description": "Wh, W, A, V, \u00b0C, \u2026",
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "timestamp",
          "connector_id",
          "measurand",
          "value",
          "unit"
        ],
        "title": "MeterValueSample",
        "type": "object"
      },
      "MeterValuesResponse": {
        "description": "`GET /api/v1/charge-points/{cp_id}/meter-values?from=\u2026&to=\u2026`.",
        "properties": {
          "cp_id": {
            "title": "Cp Id",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "samples": {
            "items": {
              "$ref": "#/components/schemas/MeterValueSample"
            },
            "title": "Samples",
            "type": "array"
          }
        },
        "required": [
          "cp_id",
          "samples",
          "request_id"
        ],
        "title": "MeterValuesResponse",
        "type": "object"
      },
...
      "Reservation": {
        "properties": {
          "connector_id": {
            "title": "Connector Id",
            "type": "integer"
          },
          "expiry_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Date"
          },
          "id_tag": {
            "title": "Id Tag",
            "type": "string"
          },
          "reservation_id": {
            "title": "Reservation Id",
            "type": "integer"
          },
          "status": {
            "description": "`active` or `cancelled`.",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "reservation_id",
          "connector_id",
          "id_tag",
          "status"
        ],
        "title": "Reservation",
        "type": "object"
      },
...
      "StatusEvent": {
        "properties": {
          "connector_id": {
            "title": "Connector Id",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "info": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Info"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "vendor_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Error Code"
          },
          "vendor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Id"
          }
        },
        "required": [
          "timestamp",
          "connector_id",
          "status"
        ],
        "title": "StatusEvent",
        "type": "object"
      },
      "StatusHistoryResponse": {
        "description": "`GET /api/v1/charge-points/{cp_id}/status-history?from=\u2026&to=\u2026`.",
        "properties": {
          "cp_id": {
            "title": "Cp Id",
            "type": "string"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/StatusEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "cp_id",
          "events",
          "request_id"
        ],
        "title": "StatusHistoryResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Inbound REST surface (E3-7 / E3-8). Backend services read charger and transaction state under `/api/v1/charge-points`, `/api/v1/transactions`, etc., and dispatch OCPP commands via `/api/v1/charge-points/{cp_id}/commands/*`. All endpoints require a bearer token (`EVEYS_OCPP_REST_INBOUND_TOKENS`); `/api/v1/health` is exempt. Errors carry the closed-enum envelope `{error, error_code, request_id}` documented in [`docs/integration/02-gateway-rest-api.md`](https://github.com/eveys-mobility/OCPP/blob/main/docs/integration/02-gateway-rest-api.md).",
    "title": "eveys/ocpp gateway",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/charge-points/{cp_id}": {
      "get": {
        "operationId": "get_charge_point_route_api_v1_charge_points__cp_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "cp_id",
            "required": true,
            "schema": {
              "title": "Cp Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChargePointDetail",
                  "additionalProperties": true,
                  "title": "Response Get Charge Point Route Api V1 Charge Points  Cp Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown cp_id (never sent BootNotification)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Charge-point detail with active reservations + charging profiles",
        "tags": [
          "charge_points"
        ]
      }
    },
...
    "/api/v1/charge-points/{cp_id}/meter-values": {
      "get": {
        "operationId": "list_meter_values_api_v1_charge_points__cp_id__meter_values_get",
        "parameters": [
          {
            "in": "path",
            "name": "cp_id",
            "required": true,
            "schema": {
              "title": "Cp Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "connector_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Connector Id"
            }
          },
          {
            "in": "query",
            "name": "measurand",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Measurand"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 10000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterValuesResponse",
                  "additionalProperties": true,
                  "title": "Response List Meter Values Api V1 Charge Points  Cp Id  Meter Values Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad from/to or window too large."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown cp_id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "ClickHouse-backed MeterValues time-series for a charge point",
        "tags": [
          "timeseries"
        ]
      }
    },
...
    "/api/v1/charge-points/{cp_id}/status-history": {
      "get": {
        "operationId": "list_status_history_api_v1_charge_points__cp_id__status_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "cp_id",
            "required": true,
            "schema": {
              "title": "Cp Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "connector_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Connector Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 10000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusHistoryResponse",
                  "additionalProperties": true,
                  "title": "Response List Status History Api V1 Charge Points  Cp Id  Status History Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad from/to or window too large."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown cp_id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "ClickHouse-backed StatusNotification history for a charge point",
        "tags": [
          "timeseries"
        ]
      }
    },
...
  }
}